Use different syntax for environment variables
Ensure that the correct version number is retrieved and used when
configuring PHP.

For my own site, I was using the same syntax but today, builds were
failing as Actions started using PHP 8 due to an error when setting up
PHP in it's step.

After switching to this syntax, the builds are now running correctly
using PHP 7.4.